gathering-dach/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
google-api-python-client==2.159.0
google-auth==2.38.0
